+++ src/monster.cpp
@@ -29,7 +29,7 @@
 #include "utils/tostring.h"
 
 
-Monster::Monster(Uint32 id, Uint16 job, Map *map):
+Monster::Monster(Uint16 id, Uint16 job, Map *map):
     Being(id, job, map)
 {
     mSprites[BASE_SPRITE] = new AnimatedSprite("graphics/sprites/monster" + toString(job - 1002) + ".xml", 0);
