+++ src/npc.cpp
@@ -33,7 +33,7 @@ extern Spriteset *npcset;
 
 NPC *current_npc = 0;
 
-NPC::NPC(Uint32 id, Uint16 job, Map *map):
+NPC::NPC(Uint16 id, Uint16 job, Map *map):
     Being(id, job, map)
 {
     mSprites[BASE_SPRITE] = new AnimatedSprite("graphics/sprites/npc.xml", job-100);
