+++ src/npc.h
@@ -29,7 +29,7 @@
 class NPC : public Being
 {
     public:
-        NPC(Uint32 id, Uint16 job, Map *map);
+        NPC(Uint16 id, Uint16 job, Map *map);
 
         virtual Type getType() const;
