+++ src/player.cpp
@@ -32,7 +32,7 @@
 
 #include "gui/gui.h"
 
-Player::Player(Uint32 id, Uint16 job, Map *map):
+Player::Player(Uint16 id, Uint16 job, Map *map):
     Being(id, job, map)
 {
     /* Load the weapon sprite. When there are more different weapons this
